Request quiz generation

+++ import_videos.py
@@ -16,9 +16,15 @@ import sqlite3
 from dotenv import load_dotenv
 from ms_client.client import MediaServerClient
 import argparse
+import logging
 
-from aristote import request_enrichment, get_enrichment, get_latest_enrichment_version
-from ubicast import handle_enrichment
+from aristote import (
+    request_enrichment,
+    get_enrichment,
+    get_latest_enrichment_version,
+    request_new_enrichment,
+)
+from ubicast import handle_enrichment, logger
 
 load_dotenv(".env")
 
@@ -29,8 +35,7 @@ CONFIG_FILE = os.environ["CONFIG_FILE"]
 def get_channel_videos(msc, oid, info=None):
     if info is None:
         info = dict(channels=0, video_oids=[])
-    print("//// Channel %s" % oid)
-    print("Making request on channels/content/ (parent_oid=%s)" % oid)
+    logger.debug("Making request on channels/content/ (parent_oid=%s)" % oid)
     response = msc.api("channels/content/", params=dict(parent_oid=oid, content="cvlp"))
     if response.get("channels"):
         for item in response["channels"]:
@@ -38,10 +43,16 @@ def get_channel_videos(msc, oid, info=None):
             get_channel_videos(msc, item["oid"], info)
     if response.get("videos"):
         for item in response["videos"]:
-            print("// Media %s" % item["oid"])
+            logger.debug("Media %s" % item["oid"])
             info["video_oids"].append(
-                dict(oid=item["oid"], parent_oid=oid, type=item["type"])
+                dict(
+                    oid=item["oid"],
+                    parent_oid=oid,
+                    type=item["type"],
+                    slug=item["slug"],
+                )
             )
+            break
     return info
 
 
@@ -68,11 +79,26 @@ def initiate_database():
         request_sent_at DATETIME,
         enrichment_notification_received_at DATETIME,
         language TEXT,
-        status TEXT
+        status TEXT,
+        name TEXT,
+        parent_oid TEXT
     )
     """
     )
 
+    conn.commit()
+
+
+def update_status_by_oid(oid: str, status: str):
+    cursor = conn.cursor()
+    cursor.execute(
+        """
+        UPDATE enrichment_requests
+        SET status = ?
+        WHERE oid = ?
+    """,
+        (status, oid),
+    )
     conn.commit()
 
 
@@ -98,7 +124,7 @@ def get_enrichment_id_by_oid(oid: str) -> str | None:
     return None
 
 
-def add_line(oid: str, enrichment_id: str, language: str):
+def add_line(oid: str, enrichment_id: str, language: str, name: str, parent_oid: str):
     cursor = conn.cursor()
 
     request_sent_at = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
@@ -106,14 +132,14 @@ def add_line(oid: str, enrichment_id: str, language: str):
 
     cursor.execute(
         """
-        INSERT INTO enrichment_requests (oid, enrichment_id, request_sent_at, language, status)
-        VALUES (?, ?, ?, ?, ?)
+        INSERT INTO enrichment_requests (oid, enrichment_id, request_sent_at, language, status, name, parent_oid)
+        VALUES (?, ?, ?, ?, ?, ?, ?)
     """,
-        (oid, enrichment_id, request_sent_at, language, status),
+        (oid, enrichment_id, request_sent_at, language, status, name, parent_oid),
     )
 
     conn.commit()
-    print(f"Enrichment request with oid: {oid} has been added.")
+    logger.debug(f"Enrichment request with oid: {oid} has been added.")
 
 
 def delete_line(oid: str):
@@ -128,7 +154,7 @@ def delete_line(oid: str):
 
     conn.commit()
 
-    print(f"Enrichment request with oid: {oid} has been deleted.")
+    logger.debug(f"Enrichment request with oid: {oid} has been deleted.")
 
 
 def oid_exists(oid):
@@ -153,19 +179,40 @@ def get_channel_language(channel_oid: str) -> str:
 def worklow(
     msc: MediaServerClient, channel_oid: str, update: str = None, limit: int = None
 ):
+    global videos_count
     global enrichment_requests_count
+    global stuck_videos
+    global enriched_videos
+
     info = get_channel_videos(msc, channel_oid)
-    stuck_videos = []
+    videos_count += len(info["video_oids"])
 
     for video in info["video_oids"]:
         if limit and enrichment_requests_count >= limit:
             break
 
         oid = video["oid"]
+        parent_oid = video["parent_oid"]
+        name = video["slug"]
+
         oid_already_exists = oid_exists(oid)
 
-        stuck = False
+        if update == "quiz" and oid_already_exists:
+            known_status = get_status_by_oid(oid)
+            if known_status == "SUCCESS":
+                enrichment_id = get_enrichment_id_by_oid(oid)
+                latest_enrichment_version = get_latest_enrichment_version(enrichment_id)
+                if latest_enrichment_version["enrichmentVersionMetadata"] is None:
+                    update_status_by_oid(oid, "PENDING")
+                    request_new_enrichment(
+                        enrichment_id, latest_enrichment_version["language"]
+                    )
+                    logger.debug(
+                        f"OID : {oid} | Enrichment : {enrichment_id} Requested quiz generation"
+                    )
+                    enriched_videos.append({"oid": oid, "enrichmentId": enrichment_id})
 
+        stuck = False
         if update == "stuck" and oid_already_exists:
             known_status = get_status_by_oid(oid)
             if known_status in ["PENDING", "FAILURE", "TRANSCRIBED"]:
@@ -185,12 +232,15 @@ def worklow(
                 if status == "FAILURE" or (
                     status == "UPLOADING_MEDIA" and ancient_upload
                 ):
+                    logger.debug(f"OID : {oid} | Enrichment : {enrichment_id} is stuck")
                     stuck = True
                     stuck_videos.append(
                         {"oid": oid, "enrichmentId": enrichment_id, "status": status}
                     )
-                    print(f"OID : {oid} | Enrichment : {enrichment_id} is stuck")
-                elif status in ["SUCCESS"]:
+                elif status == "SUCCESS":
+                    logger.debug(
+                        f"OID : {oid} | Enrichment : {enrichment_id} has been treated but missed webhook"
+                    )
                     latest_enrichment_version = get_latest_enrichment_version(
                         enrichment_id
                     )["id"]
@@ -210,18 +260,27 @@ def worklow(
                 if channel_language != "" and channel_language != "fr/en"
                 else None
             )
-            if not oid_already_exists or force_update:
+
+            ignore_video = False
+
+            if oid_already_exists:
+                known_status = get_status_by_oid(oid)
+                ignore_video = known_status in [
+                    "TRANSCRIBED_NO_LANGUAGE",
+                    "NOT_DOWNLOADABLE",
+                ]
+
+            if not oid_already_exists or (
+                oid_already_exists and force_update and not ignore_video
+            ):
                 enrichment_id = request_enrichment(oid, language=channel_language)
                 enrichment_requests_count += 1
 
-            if oid_already_exists and force_update:
+            if oid_already_exists and force_update and not ignore_video:
                 delete_line(oid)
 
-            add_line(oid, enrichment_id, channel_language)
-
-    if update == "stuck":
-        print(f"Number of stuck videos : {len(stuck_videos)}")
-        print(stuck_videos)
+            if not ignore_video:
+                add_line(oid, enrichment_id, channel_language, name, parent_oid)
 
 
 if __name__ == "__main__":
@@ -230,7 +289,7 @@ if __name__ == "__main__":
     parser.add_argument(
         "--update",
         type=str,
-        help="Specify if to force update all videos or only those stuck 'all' or 'stuck'",
+        help="Specify if to force update all videos or only those stuck 'all' or 'stuck' or 'quiz'",
     )
     parser.add_argument(
         "--csv", type=str, help="Specify a CSV file if multiple channels to treat"
@@ -238,17 +297,20 @@ if __name__ == "__main__":
     parser.add_argument(
         "--limit", type=str, help="Specify a limit for enrichment requests"
     )
+    parser.add_argument("--debug", action="store_true", help="Debug mode")
     args = parser.parse_args()
 
     channel_oid = args.channel
     update = args.update
     csv_file = args.csv
     limit = int(args.limit) if args.limit else None
+    if args.debug:
+        logger.setLevel(logging.DEBUG)
 
-    print(f"Channel: {channel_oid}")
-    print(f"Update: {update}")
-    print(f"CSV File: {csv_file}")
-    print(f"Limit : {limit}")
+    logger.info(f"Channel: {channel_oid}")
+    logger.info(f"Update: {update}")
+    logger.info(f"CSV File: {csv_file}")
+    logger.info(f"Limit : {limit}")
 
     msc = MediaServerClient(CONFIG_FILE)
     msc.check_server()
@@ -256,7 +318,10 @@ if __name__ == "__main__":
     conn = sqlite3.connect(DATABASE_URL)
     initiate_database()
 
+    videos_count = 0
     enrichment_requests_count = 0
+    stuck_videos = []
+    enriched_videos = []
 
     if channel_oid:
         worklow(msc, channel_oid, update, limit)
@@ -265,5 +330,15 @@ if __name__ == "__main__":
             reader = csv.DictReader(file)
             for row in reader:
                 worklow(msc, row["channel_oid"], update, limit)
+
+    logger.info(f"Total number of videos : {videos_count}")
+
+    if update == "stuck":
+        logger.info(f"Number of stuck videos : {len(stuck_videos)}")
+        logger.info(stuck_videos)
+
+    if update == "quiz":
+        logger.info(f"Requested enrichment for {len(enriched_videos)} videos")
+        logger.info(enriched_videos)
 
     conn.close()
